+++ OpenAI-DotNet-Proxy/Proxy/AbstractAuthenticationFilter.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
+using System.Threading.Tasks;
 
 namespace OpenAI.Proxy
 {
